feat: refactor AdminDashboardSection and add AdminDashboardSummaryCards for improved dashboard summary display
